verify path is not file

+++ replace_spaces.py
@@ -2,9 +2,10 @@ import os
 base = os.getcwd()
 for patient in os.listdir("."):
     patient = os.path.join(base, patient)
-    for dir in os.listdir(patient):
-        dir = os.path.join(patient, dir)
-        for f in os.listdir(dir):
-            r = f.replace(" ", "_")
-            if(r != f):
-                os.rename(f,r)
+    if not os.path.isfile(patient):
+        for dir in os.listdir(patient):
+            dir = os.path.join(patient, dir)
+            for f in os.listdir(dir):
+                r = f.replace(" ", "_")
+                if(r != f):
+                    os.rename(f,r)
